go/oasis-test-runner: Reorganize scenarios and helper methods
Group similar runtime scenario methods and make them globally available.
