+++ go/oasis-test-runner/scenario/e2e/helpers_network.go
@@ -1,220 +1,18 @@
-// Package e2e implements the Oasis e2e test scenarios.
 package e2e
 
 import (
 	"bytes"
-	"context"
-	"crypto/sha256"
 	"encoding/json"
 	"fmt"
-	goHash "hash"
-	"io"
 	"os"
 	"path/filepath"
 
-	flag "github.com/spf13/pflag"
-
-	"github.com/oasisprotocol/oasis-core/go/common/crypto/signature"
-	"github.com/oasisprotocol/oasis-core/go/common/entity"
-	"github.com/oasisprotocol/oasis-core/go/common/logging"
-	consensus "github.com/oasisprotocol/oasis-core/go/consensus/api"
-	"github.com/oasisprotocol/oasis-core/go/consensus/api/transaction"
-	consensusGenesis "github.com/oasisprotocol/oasis-core/go/consensus/genesis"
 	genesis "github.com/oasisprotocol/oasis-core/go/genesis/api"
 	genesisFile "github.com/oasisprotocol/oasis-core/go/genesis/file"
-	"github.com/oasisprotocol/oasis-core/go/oasis-test-runner/cmd"
 	"github.com/oasisprotocol/oasis-core/go/oasis-test-runner/env"
 	"github.com/oasisprotocol/oasis-core/go/oasis-test-runner/oasis"
 	"github.com/oasisprotocol/oasis-core/go/oasis-test-runner/oasis/cli"
-	"github.com/oasisprotocol/oasis-core/go/oasis-test-runner/scenario"
-	registry "github.com/oasisprotocol/oasis-core/go/registry/api"
-	staking "github.com/oasisprotocol/oasis-core/go/staking/api"
 )
-
-const (
-	// cfgNodeBinary is the path to oasis-node executable.
-	cfgNodeBinary = "node.binary"
-)
-
-// ParamsDummyScenario is a dummy instance of E2E scenario used to register global E2E flags.
-var ParamsDummyScenario = NewScenario("")
-
-// Scenario is a base scenario for oasis-node end-to-end tests.
-type Scenario struct {
-	Net    *oasis.Network
-	Flags  *env.ParameterFlagSet
-	Logger *logging.Logger
-
-	name string
-}
-
-// NewScenario creates a new base scenario for oasis-node end-to-end tests.
-func NewScenario(name string) *Scenario {
-	// Empty scenario name is used for registering global parameters only.
-	fullName := "e2e"
-	if name != "" {
-		fullName += "/" + name
-	}
-
-	sc := &Scenario{
-		name:   fullName,
-		Logger: logging.GetLogger("scenario/" + fullName),
-		Flags:  env.NewParameterFlagSet(fullName, flag.ContinueOnError),
-	}
-	sc.Flags.String(cfgNodeBinary, "oasis-node", "path to the node binary")
-
-	return sc
-}
-
-// Clone implements scenario.Scenario.
-func (sc *Scenario) Clone() Scenario {
-	return Scenario{
-		Net:    sc.Net,
-		Flags:  sc.Flags.Clone(),
-		Logger: sc.Logger,
-		name:   sc.name,
-	}
-}
-
-// Name implements scenario.Scenario.
-func (sc *Scenario) Name() string {
-	return sc.name
-}
-
-// Parameters implements scenario.Scenario.
-func (sc *Scenario) Parameters() *env.ParameterFlagSet {
-	return sc.Flags
-}
-
-// PreInit implements scenario.Scenario.
-func (sc *Scenario) PreInit(childEnv *env.Env) error {
-	return nil
-}
-
-// Fixture implements scenario.Scenario.
-func (sc *Scenario) Fixture() (*oasis.NetworkFixture, error) {
-	nodeBinary, _ := sc.Flags.GetString(cfgNodeBinary)
-
-	return &oasis.NetworkFixture{
-		Network: oasis.NetworkCfg{
-			NodeBinary: nodeBinary,
-			Consensus: consensusGenesis.Genesis{
-				Parameters: consensusGenesis.Parameters{
-					GasCosts: transaction.Costs{
-						consensusGenesis.GasOpTxByte: 1,
-					},
-				},
-			},
-		},
-		Entities: []oasis.EntityCfg{
-			{IsDebugTestEntity: true},
-			{},
-		},
-		Validators: []oasis.ValidatorFixture{
-			{Entity: 1, Consensus: oasis.ConsensusFixture{SupplementarySanityInterval: 1}},
-			{Entity: 1},
-			{Entity: 1},
-		},
-		Seeds: []oasis.SeedFixture{{}},
-	}, nil
-}
-
-// Init implements scenario.Scenario.
-func (sc *Scenario) Init(childEnv *env.Env, net *oasis.Network) error {
-	sc.Net = net
-	return nil
-}
-
-// GetExportedGenesisFiles gathers exported genesis files and ensures
-// all exported genesis files match.
-func (sc *Scenario) GetExportedGenesisFiles(skipCompute bool) ([]string, error) {
-	dumpGlob := "genesis-*.json"
-
-	// Gather all nodes.
-	var nodes []interface {
-		ExportsPath() string
-	}
-	for _, v := range sc.Net.Validators() {
-		nodes = append(nodes, v)
-	}
-	if !skipCompute {
-		for _, n := range sc.Net.ComputeWorkers() {
-			nodes = append(nodes, n)
-		}
-	}
-	for _, n := range sc.Net.Keymanagers() {
-		nodes = append(nodes, n)
-	}
-
-	// Gather all genesis files.
-	var files []string
-	for _, node := range nodes {
-		dumpGlobPath := filepath.Join(node.ExportsPath(), dumpGlob)
-		globMatch, err := filepath.Glob(dumpGlobPath)
-		if err != nil {
-			return nil, fmt.Errorf("glob failed: %s: %w", dumpGlobPath, err)
-		}
-		if len(globMatch) == 0 {
-			return nil, fmt.Errorf("genesis file not found in: %s", dumpGlobPath)
-		}
-		if len(globMatch) > 1 {
-			return nil, fmt.Errorf("more than one genesis file found in: %s", dumpGlobPath)
-		}
-		files = append(files, globMatch[0])
-	}
-
-	// Assert all exported files match.
-	var firstHash goHash.Hash
-	for _, file := range files {
-		// Compute hash.
-		f, err := os.Open(file)
-		if err != nil {
-			return nil, fmt.Errorf("failed to open file: %s: %w", file, err)
-		}
-		defer f.Close()
-		hnew := sha256.New()
-		if _, err := io.Copy(hnew, f); err != nil {
-			return nil, fmt.Errorf("sha256 failed on: %s: %w", file, err)
-		}
-		if firstHash == nil {
-			firstHash = hnew
-		}
-
-		// Compare hash with first hash.
-		if !bytes.Equal(firstHash.Sum(nil), hnew.Sum(nil)) {
-			return nil, fmt.Errorf("exported genesis files do not match %s, %s", files[0], file)
-		}
-	}
-
-	return files, nil
-}
-
-func (sc *Scenario) GetTestEntityNonce(ctx context.Context) (uint64, error) {
-	ent, _, err := entity.TestEntity()
-	if err != nil {
-		return 0, err
-	}
-	return sc.GetEntityNonce(ctx, ent)
-}
-
-func (sc *Scenario) GetEntityNonce(ctx context.Context, ent *entity.Entity) (uint64, error) {
-	addr := staking.NewAddress(ent.ID)
-	return sc.Net.ClientController().Consensus.GetSignerNonce(ctx, &consensus.GetSignerNonceRequest{
-		Height:         consensus.HeightLatest,
-		AccountAddress: addr,
-	})
-}
-
-func (sc *Scenario) GetEntityNonceByID(ctx context.Context, id signature.PublicKey) (uint64, error) {
-	ent, err := sc.Net.ClientController().Registry.GetEntity(ctx, &registry.IDQuery{
-		Height: consensus.HeightLatest,
-		ID:     id,
-	})
-	if err != nil {
-		return 0, err
-	}
-	return sc.GetEntityNonce(ctx, ent)
-}
 
 // Flag for consensus state reset.
 const (
@@ -475,58 +273,4 @@ func (sc *Scenario) finishWithoutChild() error {
 	default:
 		return sc.Net.CheckLogWatchers()
 	}
-}
-
-// RegisterScenarios registers all end-to-end scenarios.
-func RegisterScenarios() error {
-	// Register non-scenario-specific parameters.
-	cmd.RegisterScenarioParams(ParamsDummyScenario.Name(), ParamsDummyScenario.Parameters())
-
-	// Register default scenarios which are executed, if no test names provided.
-	for _, s := range []scenario.Scenario{
-		// Registry CLI test.
-		RegistryCLI,
-		// Stake CLI test.
-		StakeCLI,
-		// Gas fees tests.
-		GasFeesStaking,
-		GasFeesStakingDumpRestore,
-		// Identity CLI test.
-		IdentityCLI,
-		// Genesis file test.
-		GenesisFile,
-		// Node upgrade tests.
-		NodeUpgradeDummy,
-		NodeUpgradeMaxAllowances,
-		NodeUpgradeV62,
-		NodeUpgradeEmpty,
-		NodeUpgradeCancel,
-		// Debonding entries from genesis test.
-		Debond,
-		// Early query test.
-		EarlyQuery,
-		EarlyQueryInitHeight,
-		// Consensus state sync.
-		ConsensusStateSync,
-		// Multiple seeds test.
-		MultipleSeeds,
-		// Seed API test.
-		SeedAPI,
-		// ValidatorEquivocation test.
-		ValidatorEquivocation,
-		// Byzantine VRF beacon tests.
-		ByzantineVRFBeaconHonest,
-		ByzantineVRFBeaconEarly,
-		ByzantineVRFBeaconMissing,
-		// Minimum transact balance test.
-		MinTransactBalance,
-		// Consensus governance update parameters tests.
-		ChangeParametersMinCommissionRate,
-	} {
-		if err := cmd.Register(s); err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
