+++ go/oasis-test-runner/scenario/e2e/runtime/storage_sync_from_registered.go
@@ -93,7 +93,7 @@ func (sc *storageSyncFromRegisteredImpl) Run(ctx context.Context, childEnv *env.
 	nextEpoch++
 
 	// Wait for the client to exit.
-	if err = sc.WaitTestClientOnly(); err != nil {
+	if err = sc.WaitTestClient(); err != nil {
 		return err
 	}
 
@@ -183,8 +183,5 @@ func (sc *storageSyncFromRegisteredImpl) Run(ctx context.Context, childEnv *env.
 	// Run the client again.
 	sc.Logger.Info("starting a second client to check if runtime works with compute worker 1")
 	sc.Scenario.testClient = NewKVTestClient().WithSeed("seed2").WithScenario(InsertRemoveKeyValueEncScenarioV2)
-	if err = sc.startTestClientOnly(ctx, childEnv); err != nil {
-		return err
-	}
-	return sc.waitTestClient()
+	return sc.RunTestClientAndCheckLogs(ctx, childEnv)
 }
